fix: tests no longer try to call aws api's (#35)
Move mock from aws sdk to the openidconnect code. Mock authorize &
getsecret functions. Remove test that was testing the oidc-code instead
of the auth function.

Adds the option to inject an OpenIDConnect object into the handler,
we use this to provide one with mocked methods, so the application
doesn't try to call aws api's while testing.